chore: remove unused code and add more comments

+++ src/graph/graph.cpp
@@ -2,14 +2,14 @@
 
 #include "graph.h"
 
-graph::graph(ushort vertices, ushort edges) {
+graph::graph(ushort vertices, ulong edges) {
     this->vertices = vertices;
     this->edges = edges;
 }
 
-graph::graph(ushort vertices, double fill, bool is_directed): graph(vertices,static_cast<ushort>(vertices * (vertices - 1) * fill / (is_directed ? 2 : 1))) {}
+graph::graph(ushort vertices, double fill, bool is_directed): graph(vertices, static_cast<ulong>(vertices * (vertices - 1) * fill / (is_directed ? 2 : 1))) {}
 
-ushort graph::get_edges() const {
+ulong graph::get_edges() const {
     return this->edges;
 }
